requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements.in
#
jinja2==3.1.2
    # via -r ./requirements.in
markupsafe==2.1.2
    # via jinja2
pyyaml==6.0
    # via -r ./requirements.in
